More robust identifier for apps
- Support collecting logs from apps external to geOrchestra, where they
can be served at the root of the server. For instance, a mapserver
instance served on its own subdomain, at `/` (the name of the app is not
provided in the path).
For now, only OpenTelemetry supports it. It expects that the
OpenTelemetry feed is collecting the `server.address` attribute. The app
id hence generated is a concatenation of the fqdn provided by the
server.address attribute and the path (`/` in that case). The admin has to
declare such identifiers in the config, for the mapping between the id
and the app name. It also requires the config option
`support_multiple_dn:true`.
If the Otel provider can't get the `server.address` attribute, one can
add it during the logs collection by Vector.
- Also supports collecting logs from several similar instances. Same
  config will prepend the apps paths by the FQDN, hence ensuring that we
can discriminate them.

The main drawback is that when this option is activated, all the
mappings have to be declared explicitly in the config file.